backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.5.1",
    "pillow>=11.2.1",
]
